move ActionButtons and ContentMapBox into MapCard
