+++ src/pages/AnalyzeProjectSites.jsx
@@ -131,7 +131,6 @@ export default function AnalyzeProjectSite(props) {
 
   return (
     <GenericMapHolder
-      map={map}
       isItAGraph={analyzeAreaState.isItAGraph}
       leftColumn={
         <AnalyzeProjectSiteLeftColumn
